agent/requirements.txt
kafka-python
psutil